+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Additions:
   * Add refinement Float#to_km
   * Add constant AIXM::UNLIMITED
+  * Add short_name to Airspace
 
 ## 0.1.0